Add use cases changes.
