+++ src/main/java/use_cases/timer/RunTimer.java
@@ -1,0 +1,4 @@
+package use_cases.timer;
+
+public class RunTimer {
+}
