+++ src/main/java/use_cases/timer/String_to_int.java
@@ -1,4 +1,4 @@
-package use_cases.timer_use_cases;
+package use_cases.timer;
 
 import java.util.Scanner;
