+++ src/main/java/use_cases/timer/TimerOutputBoundary.java
@@ -1,6 +1,6 @@
-package use_cases.timer_use_cases.create_single_timer;
+package use_cases.timer;
 
-public class CreateWorkTimerOutputBoundary {
+public class TimerOutputBoundary {
 
     private Boolean success;
 
@@ -10,7 +10,7 @@ public class CreateWorkTimerOutputBoundary {
      * @param w - The minutes need of work timer
      * @param r - The minutes need of work timer
      */
-    public CreateWorkTimerOutputBoundary(String w, String r){
+    public TimerOutputBoundary(String w, String r){
         if(!(w.equals("") && r.equals(""))){
             //CreateTimer.createHabit(w, r, i);
             this.success = true;
@@ -18,7 +18,7 @@ public class CreateWorkTimerOutputBoundary {
     }
 
     /**
-     * @return - Returns whether habit was created successfully.
+     * @return - Returns whether timer was created successfully.
      */
     public Boolean getSuccess() {
         return this.success;
