+++ src/main/java/use_cases/timer/create_single_timer/CreateWorkTimerOutputBoundary.java
@@ -1,6 +1,6 @@
-package use_cases.timer_use_cases;
+package use_cases.timer.create_single_timer;
 
-public class TimerOutputBoundary {
+public class CreateWorkTimerOutputBoundary {
 
     private Boolean success;
 
@@ -10,7 +10,7 @@ public class TimerOutputBoundary {
      * @param w - The minutes need of work timer
      * @param r - The minutes need of work timer
      */
-    public TimerOutputBoundary(String w, String r){
+    public CreateWorkTimerOutputBoundary(String w, String r){
         if(!(w.equals("") && r.equals(""))){
             //CreateTimer.createHabit(w, r, i);
             this.success = true;
@@ -18,7 +18,7 @@ public class TimerOutputBoundary {
     }
 
     /**
-     * @return - Returns whether timer was created successfully.
+     * @return - Returns whether habit was created successfully.
      */
     public Boolean getSuccess() {
         return this.success;
