+++ src/main/java/use_cases/timer/edit_timer/EditTimerInputBoundary.java
@@ -1,4 +1,4 @@
-package use_cases.timer_use_cases.edit_timer;
+package use_cases.timer.edit_timer;
 
 /**
  * This is an interface that passes the inputData to the interactor of timer modification/edition.
