+++ src/main/java/use_cases/timer/edit_timer/EditTimerOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_cases.timer_use_cases.edit_timer;
+package use_cases.timer.edit_timer;
 
 /**
  * This is an interface that passes the output to the Presenter of task edition/modification.
