+++ src/main/java/use_cases/timer/edit_timer/EditTimerOutputData.java
@@ -1,4 +1,4 @@
-package use_cases.timer_use_cases.edit_timer;
+package use_cases.timer.edit_timer;
 
 import javax.swing.*;
 /**
